feat: add table structure
